SCALE-SEAM (84): the cluster that split three ways, and a right number corrected to a wrong one
Twelve methods out of client.ts (1,292 -> 1,212), plus the DevBudget*
interfaces from client.ts into types.ts where a mixin can import them.

  what is this development costing,   devBudget, saveDevBudget,
  and how is it funded?               devBudgetCostLines, sourcesUses,
                                      gmpReconciliation, syncGmpToHard,
                                      constructionDraws, loanDraws,
                                      loanDrawRequestPdf -> proforma.ts
  are we priced right, and does       pricingReconcile, costTraceability
  cost tie back to the model?         -> cost.ts
  what has each subcontract billed?   subcontractorBilling -> accounting.ts

THE TYPES DECIDED THE HARDEST PLACEMENT. gmpReconciliation and
syncGmpToHard look like cost.ts work - ㉚ put the GMP/pay-app stack there
and both routes are GMP-flavoured. But cost.ts's gmpBudget is the GC's OWN
GMP (contract value, categories, EAC/ETC) while these two are the DEVELOPER
comparing their hard cost against it, and both return DevBudgetLine /
DevBudgetSummary. Filing them in cost.ts would have forced the DevBudget
type family into two mixins. A split that duplicates a type family across
mixins is almost always the wrong seam - a sharper rule than "group by what
they answer", because it is mechanical.

(83) REPORTED 44 WHERE THE ANSWER WAS 49, AND 49 WAS ALREADY WRITTEN DOWN.

The CX-1 banner really held 49 methods. (83) counted them with

    /^  ([a-zA-Z_]\w*)\(/

which does not match `async` methods, so five were invisible: importRvt,
loanDrawRequestPdf, raisePlan, takeoffDxf, uploadSourceIfc. The roadmap had
said 49. (83) "corrected" the correct number to an incorrect one and
shipped 44 into the changelog, the roadmap and its own merge commit - in a
slice whose subject was counts that disagree with the lists beside them.

The failure is not carelessness with a number. A regex that silently drops
a syntactic variant returns a smaller number, not an error, and a smaller
number looks exactly like a right one. Same shape as the inverted count
#396 pinned, the two totals #399 fixed, and a && chained off head that
printed "TYPECHECK CLEAN" over a failing tsc. Four in one day, every one a
check whose output did not measure what it claimed.

So the map is no longer proofread, it is DERIVED, by
apps/web/src/api/unfiledMap.test.ts: the declared count must equal the
methods actually below the banner, every such method must be named, and no
name may survive its method having moved. Its regex matches async. Which
tokens count as method names is itself derived - a word qualifies only if
some file in api/ declares a method by it - because a hand-kept list would
miss `property` and `templates` (no capital) or swallow prose like "check"
and "stay". All three assertions were mutation-checked: a wrong count, a
dropped name, and a name left behind after its method moved.

The gate found two more things at once. The map had never listed
takeoffDxf, raisePlan, uploadSourceIfc or importRvt - all async, all
invisible to (83)'s query - and they are a cluster it had no line for:
"how do I get a file into this project?", four multipart uploads. And
rvtBridgeStatus is the bridge importRvt drives, so (83)'s claim that it was
a client-level "stayer" was wrong; it was only plausible while the four
uploads beside it could not be seen. Four stayers remain: enumOptions,
searchAll, attachmentUrl, templates.

Verification: tsc --noEmit exit 0 (the new test failed tsc on strict-null
after vitest passed it - vitest does not typecheck, so a green suite was
not evidence the file compiled); full web suite 203 files / 2,059 tests;
eslint clean; test_file_sizes pin 1,292 -> 1,212 and exact;
test_claude_md_gates 301 citations; test_changelog_current; ruff clean.
